Fix for deadlock-y condition on error + close call (#20)
The root problem is that in our `eof` call on an `SSLStream`, we're
currently grabbing the `SSLStream` lock for basically the entire
`eof` call, _including_ when we do a potentially long-blocking call
to `eof` on the underlying socket! This is bad because it basically
prevents force-closing the `SSLStream` until the socket `eof` call
returns, and we can finally get the `SSLStream` lock in our `close` call.

The fix proposed here is introducing a 2nd `eoflock` and renaming `lock`
to `closelock`. `closelock` is used (hopefully obviously) just for
protecting the close operation. `eoflock` is used to bundle the calling
of `eof` on the underlying socket and calling `SSL_peek` to process
bytes, so they're observed by concurrent tasks as a single operation.